+++ sndata/csp/_dr3.py
@@ -249,7 +249,7 @@ class DR3(PhotometricRelease, DefaultParser):
         for file_name in self._filter_file_names:
             utils.download_file(
                 url=self._filter_url + file_name,
-                path=self._filter_dir / file_name,
+                destination=self._filter_dir / file_name,
                 force=force,
                 timeout=timeout
             )